% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bolstad2-package.R
\docType{data}
\name{c10ex16.df}
\alias{c10ex16.df}
\alias{ex16.df}
\title{Chapter 10 Example 16 data}
\format{
A data frame with 10 observations in a single variable called y
}
\description{
A random sample of size 10 from a \eqn{N(\mu, \sigma^{2})} distribution
where both mu and sigma are unknown parameters.
}
\keyword{datasets}
