% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bolstad2-package.R
\docType{data}
\name{chd.df}
\alias{chd.df}
\title{Coronary Heart Disease Chapter 8 Example 11}
\format{
A data frame with 100 observations in two columns \tabular{lrrr}{
[,1] \tab age \tab numeric \tab age in years \cr [,2] \tab chd \tab numeric
factor \tab coronary heat disease status. Levels (1 = Yes), (0 = No) \cr }
}
\description{
The age and coronory heart disease status of 100 individuals taken from
Hosmer and Lemeshow (1989).
}
\keyword{datasets}
