\name{hiermeanRegTest.df}
\docType{data}
\alias{hiermeanRegTest.df}
\title{Test data for hiermeanReg}
\description{
  Data for testing hiermeanReg which uses Gibbs sampling on a hierarchical
  normal mean model with regression on covariates
}
\usage{data(hiermeanRegTest.df)
}
\format{
  A data frame with 30 observations on 4 variables.
  \tabular{rlll}{\
    [1,] \tab y \tab numeric \tab the response vector \cr
    [2,] \tab group \tab factor \tab the grouping factor levels 1-3 \cr
    [3,] \tab x1 \tab numeric \tab the first covariate \cr
    [4,] \tab x2 \tab numeric \tab the second covariate \cr
    }
}
\seealso{hiermeanReg
}
\keyword{datasets}
