% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bolstad2-package.R
\docType{data}
\name{poissonTest.df}
\alias{poissonTest.df}
\title{A test data set for bayesPois}
\format{
A data frame with 100 observations on 5 variables.  \tabular{rlll}{
[1,] \tab x \tab numeric \tab the covariate \cr [2,] \tab eps \tab numeric
\tab the error in the log response \cr [3,] \tab log.lam \tab numeric \tab
\eqn{\log(\lambda_{i}) = 1 + 5x_{i} + \epsilon_{i}} where \eqn{\epsilon_{i}}
\cr [4,] \tab lam \tab numeric \tab \eqn{\exp(\log(\lambda))} \cr [5,] \tab
y \tab numeric \tab a Poisson random variate with mean \eqn{\lambda_{i}} \cr
}
}
\description{
A test data set for bayesPois. The data come from the equation
\eqn{\log(\lambda_{i}) = 1 + 5x_{i} + \epsilon_{i}} where \eqn{\epsilon_{i}}
comes from N(0,0.01).
}
\seealso{
bayesPois
}
\keyword{datasets}
