\name{c10ex16.df}
\alias{c10ex16.df}
\alias{ex16.df}
\docType{data}
\title{Chapter 10 Example 16 data}
\description{
  A random sample of size 10 from a \eqn{N(\mu, \sigma^{2})} distribution
  where both mu and sigma are unknown parameters.
}
\usage{data(c10ex16.df)
}
\format{
  A data frame with 10 observations in a single variable called y
  }
\keyword{datasets}
