\name{chd.df}
\alias{chd.df}
\docType{data}
\title{Coronary Heart Disease Chapter 8 Example 11}
\description{
  The age and coronory heart disease status of 100 individuals taken
  from Hosmer and Lemeshow (1989).
}
\usage{
data(chd.df)
}
\format{
  A data frame with 100 observations in two columns
  \tabular{lrrr}{
    [,1] \tab age \tab numeric \tab age in years \cr
    [,2] \tab chd \tab numeric factor \tab coronary heat disease
    status. Levels (1 = Yes), (0 = No) \cr
  }
}
\keyword{datasets}
