% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bolstad2-package.R
\docType{data}
\name{hiermeanRegTest.df}
\alias{hiermeanRegTest.df}
\title{Test data for hiermeanReg}
\format{
A data frame with 30 observations on 4 variables.  \tabular{rlll}{\
[1,] \tab y \tab numeric \tab the response vector \cr [2,] \tab group \tab
factor \tab the grouping factor levels 1-3 \cr [3,] \tab x1 \tab numeric
\tab the first covariate \cr [4,] \tab x2 \tab numeric \tab the second
covariate \cr }
}
\description{
Data for testing hiermeanReg which uses Gibbs sampling on a hierarchical
normal mean model with regression on covariates
}
\seealso{
hiermeanReg
}
\keyword{datasets}
